Implement "optional" fields
Fields are marked "optional" by adding the type "aeon:optional_field()"
as one of the possible types.

Optional fields may not appear in the JSON, in which case conversion
should not fail, and the record field left at it's default value.

If a field is optional and has 'null' or 'undefined' value in the record,
it is not included in the resulting JSON